+++ js/world.js
@@ -1,9 +1,10 @@
 import * as THREE from 'three';
-import { createTerrain, createBarriers, createTrees, createClouds, createStarterHouse } from './worldGeneration.js';
+import { createTerrain, createBarriers, createTrees, createClouds, createStarterHouse, createShopkeeper, createAmphitheatre } from './worldGeneration.js';
 
 export class World {
-    constructor(scene) {
+    constructor(scene, npcManager) {
         this.scene = scene;
+        this.npcManager = npcManager;
         this.terrain = null;
     }
 
@@ -12,7 +13,10 @@ export class World {
         createBarriers(this.scene, this.terrain.userData.getHeight);
         createTrees(this.scene, this.terrain.userData.getHeight);
         createClouds(this.scene);
-        createStarterHouse(this.scene, this.terrain.userData.getHeight);
+        const house = createStarterHouse(this.scene, this.terrain.userData.getHeight);
+        createShopkeeper(this.scene, this.terrain, this.npcManager, house.position);
+        createAmphitheatre(this.scene, this.terrain.userData.getHeight);
         return this.terrain;
     }
 }
+