+++ js/worldGeneration.js
@@ -5,3 +5,6 @@ export { createTrees } from './worldgen/trees.js';
 export { createClouds } from './worldgen/clouds.js';
 export { createStarterHouse } from './worldgen/starterHouse.js';
 export { createGroundGrid } from './worldgen/groundGrid.js';
+export { createShopkeeper } from './worldgen/shopkeeper.js';
+export { createAmphitheatre } from './worldgen/amphitheatre.js';
+